Flyttet RequestInfo til eget csproj, og forenklet klassen
