--- UnitTests/examples/withparams/endpoint/endpoint.json
@@ -1,11 +1,0 @@
-﻿{
-    "name": "Endpoint",
-    "pathregex": "^/endpoint/",
-
-    "responses": [
-        {
-            "match": {},
-            "script": "script.csscript"
-        }
-    ]
-}
