--- UnitTests/examples/withparams/endpoint/params.json
@@ -1,13 +1,0 @@
-﻿[
-    {
-        "name": "greeting",
-        "default": "Hello",
-        "description": "The greeting to display"
-    },
-
-    {
-        "name": "subject",
-        "default": "World",
-        "description": "Recipient of greeting"
-    }
-]
