--- UnitTests/examples/withparams/endpoint/script.csscript
@@ -1,1 +1,0 @@
-﻿return GetParam("greeting") + " " + GetParam("subject");
