+++ netmockery/netmockery.csproj
@@ -17,4 +17,7 @@
     <PackageReference Include="System.Xml.XPath" Version="4.3.0" />
     <PackageReference Include="System.Xml.XPath.XDocument" Version="4.3.0" />
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\netmockery.globals\netmockery.globals.csproj" />
+  </ItemGroup>
 </Project>
